+++ mooc-portal/src/main/java/cn/edu/lingnan/mooc/portal/controller/CourseController.java
@@ -13,74 +13,79 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/courses")
 public class CourseController {
+
     @Autowired
     private CourseService receptionCourseService;
 
     /**
-     * 前台方法
      * 根据标签查询课程
+     *
      * @param queryCourseParam
      * @return
      */
     @PostMapping("/getByTag")
-    public RespResult findCourseByTagIdList2(@RequestBody QueryCourseParam queryCourseParam){
-        if(queryCourseParam.getPageIndex() == null && queryCourseParam.getPageIndex() < 1){
+    public RespResult findCourseByTagIdList2(@RequestBody QueryCourseParam queryCourseParam) {
+        if (queryCourseParam.getPageIndex() == null && queryCourseParam.getPageIndex() < 1) {
             queryCourseParam.setPageIndex(1);
         }
-        if(queryCourseParam.getPageSize() == null){
+        if (queryCourseParam.getPageSize() == null) {
             queryCourseParam.setPageSize(15);
         }
         return RespResult.success(receptionCourseService.getCourseByTagList(queryCourseParam.getTagIdList(),
-                queryCourseParam.getPageIndex(),queryCourseParam.getPageSize()));
+                queryCourseParam.getPageIndex(), queryCourseParam.getPageSize()));
     }
 
     /**
      * 根据课程Id查询课程基本信息
+     *
      * @param id
      * @return
      */
     @GetMapping("/{id}")
-    public RespResult findCourseById(@PathVariable Integer id){
+    public RespResult findCourseById(@PathVariable Long id) {
         return RespResult.success(receptionCourseService.findCourseDetailById(id));
     }
 
     /**
-     * 根据教师id查询 她/他 的开设的课程
+     * 根据教师id查询开设的课程
+     *
      * @param userId
      * @param pageIndex
      * @param pageSize
      * @return
      */
     @GetMapping("/listByTeachId")
-    public RespResult findCourseByTeachId(@RequestParam(value = "userId") Integer userId,
-                                 @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
-                                 @RequestParam(value = "pageSize", defaultValue = "15") Integer pageSize) {
+    public RespResult findCourseByTeachId(@RequestParam(value = "userId") Long userId,
+                                          @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
+                                          @RequestParam(value = "pageSize", defaultValue = "15") Integer pageSize) {
 
         return RespResult.success(receptionCourseService.findCourseByTeachId(userId, pageIndex, pageSize));
     }
 
 
     /**
-     * 根据用户id查询 她/他 的收藏的课程
+     * 根据用户id查询收藏的课程
+     *
      * @param userId
      * @param pageIndex
      * @param pageSize
      * @return
      */
     @GetMapping("/collection/list")
-    public RespResult findCollectionCourseByUserId(@RequestParam(value = "userId") Integer userId,
-                                          @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
-                                          @RequestParam(value = "pageSize", defaultValue = "15") Integer pageSize) {
+    public RespResult findCollectionCourseByUserId(@RequestParam(value = "userId") Long userId,
+                                                   @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
+                                                   @RequestParam(value = "pageSize", defaultValue = "15") Integer pageSize) {
 
         return RespResult.success(receptionCourseService.findCollectionCourseByUserId(userId, pageIndex, pageSize));
     }
 
     /**
      * 点击收藏或者取消收藏
+     *
      * @return
      */
     @PutMapping("/collectionOrCancel/{courseId}")
-    public RespResult collectionOrCancel(@PathVariable Integer courseId){
+    public RespResult collectionOrCancel(@PathVariable Long courseId) {
         receptionCourseService.collectionOrCancel(courseId);
         return RespResult.success();
     }
