+++ mooc-portal/src/main/java/cn/edu/lingnan/mooc/portal/model/vo/UserDetailVO.java
@@ -11,18 +11,26 @@ import lombok.ToString;
 @ToString
 public class UserDetailVO {
 
-    private Integer id;
-    // 用户头像
+    private Long id;
+    /**
+     * 用户头像
+     */
     private String userImage;
-    // 用户昵称
+    /**
+     * 用户昵称
+     */
     private String name;
-    // 登录账号
+    /**
+     * 登录账号
+     */
     private String account;
     /**
      * 个人座右铭、格言
      */
     private String motto;
-    // 类型，教师/普通用户
+    /**
+     * 类型，教师/普通用户
+     */
     private String userType;
 
 }
